flatMapConcat impl. -WIP

+++ akka-stream/src/main/scala/akka/stream/ContextMapStrategy.scala
@@ -7,21 +7,21 @@ package akka.stream
 object ContextMapStrategy {
 
   final case class Iterate[In, Ctx, Out](
-      iterateFn: (In, Ctx, Out, Int, Boolean) => Ctx,
+      iterateFn: (In, Ctx, Out, Long, Boolean) => Ctx,
       emptyFn: Option[(In, Ctx) => (Out, Ctx)] = None)
 
   /**
    * Passthrough the same context [[Ctx] of the input element [[In]] to all output elements [[Out]].
    */
   def same[In, Ctx, Out](): Iterate[In, Ctx, Out] = Iterate(
-    iterateFn = (_: In, inCtx: Ctx, _: Out, _: Int, _: Boolean) => inCtx
+    iterateFn = (_: In, inCtx: Ctx, _: Out, _: Long, _: Boolean) => inCtx
   )
 
   /**
    * Transform the context of the first element given the input element [[In]] and context [[Ctx]].
    */
   def first[In, Ctx, Out](firstFn: (In, Ctx, Out) => Ctx): Iterate[In, Ctx, Out] = Iterate(
-    iterateFn = (in: In, inCtx: Ctx, out: Out, index: Int, _: Boolean) =>
+    iterateFn = (in: In, inCtx: Ctx, out: Out, index: Long, _: Boolean) =>
       if (index == 0) firstFn(in, inCtx, out)
       else inCtx
   )
@@ -30,8 +30,8 @@ object ContextMapStrategy {
    * Transform the context of the last element given the input element [[In]], context [[Ctx]], and the index
    * of the element.
    */
-  def last[In, Ctx, Out](fn: (In, Ctx, Out, Int) => Ctx): Iterate[In, Ctx, Out] = Iterate(
-    iterateFn = (in: In, inCtx: Ctx, out: Out, index: Int, hasNext: Boolean) =>
+  def last[In, Ctx, Out](fn: (In, Ctx, Out, Long) => Ctx): Iterate[In, Ctx, Out] = Iterate(
+    iterateFn = (in: In, inCtx: Ctx, out: Out, index: Long, hasNext: Boolean) =>
       if (!hasNext) fn(in, inCtx, out, index)
       else inCtx
   )
@@ -41,7 +41,7 @@ object ContextMapStrategy {
    * output element [[Out]], index, and if the there is a next output element [[Out]]. This strategy can be used to
    * satisfy all 1:many use cases.
    */
-  def iterate[In, Ctx, Out](fn: (In, Ctx, Out, Int, Boolean) => Ctx): Iterate[In, Ctx, Out] = Iterate(fn)
+  def iterate[In, Ctx, Out](fn: (In, Ctx, Out, Long, Boolean) => Ctx): Iterate[In, Ctx, Out] = Iterate(fn)
 
   /**
    * Iterate over each element and transform the context given the element [[In]], context [[InCtx]],
@@ -50,7 +50,7 @@ object ContextMapStrategy {
    * This strategy can be used to satisfy all 1:many and 1:0 use cases.
    */
   def iterateOrEmpty[In, Ctx, Out](
-                                    fn: (In, Ctx, Out, Int, Boolean) => Ctx,
+                                    fn: (In, Ctx, Out, Long, Boolean) => Ctx,
                                     emptyFn: (In, Ctx) => (Out, Ctx)
                                   ): Iterate[In, Ctx, Out] = Iterate(iterateFn = fn, emptyFn = Some(emptyFn))
 }
